[REF] split beesdoo_base -> eater
[UPD] Update eater.pot
