+++ beesdoo_base/views/partner.xml
@@ -2,26 +2,11 @@
 <odoo>
 
 
-    <record model="ir.actions.act_window" id="action_eater_wizard">
-        <field name="name">New Eater</field>
-        <field name="res_model">eater.new.wizard</field>
-        <field name="view_mode">form</field>
-        <field name="target">new</field>
-    </record>
-
     <record model="ir.ui.view" id="beesdoo_partner_form_view">
         <field name="name">beesdoo.partner.form.view</field>
         <field name="model">res.partner</field>
         <field name="inherit_id" ref="base.view_partner_form" />
         <field name="arch" type="xml">
-            <field name="website" position="after">
-                <field name="eater" attrs="{'invisible': [('customer', '=', False)]}" />
-                <field
-                    name="parent_eater_id"
-                    attrs="{'invisible' : [('eater', '!=', 'eater')]}"
-                />
-            </field>
-
 
             <!-- Temporary fix for the partner_view -->
             <xpath expr="//field[@name='customer']" position='before'>
