+++ beesdoo_base/wizard/__init__.py
@@ -1,2 +1,1 @@
-from . import partner
 from . import portal_wizard
