+++ eater/i18n/eater.pot
@@ -1,0 +1,57 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* eater
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: eater
+#: model_terms:ir.ui.view,arch_db:eater.Eater Wizard
+msgid "Cancel"
+msgstr ""
+
+#. module: eater
+#: model:ir.model,name:eater.model_res_partner
+msgid "Contact"
+msgstr ""
+
+#. module: eater
+#: model:ir.model.fields,field_description:eater.field_res_partner__country_id
+#: model:ir.model.fields,field_description:eater.field_res_users__country_id
+msgid "Country"
+msgstr ""
+
+#. module: eater
+#: model_terms:ir.ui.view,arch_db:eater.Eater Wizard
+msgid "Create"
+msgstr ""
+
+#. module: eater
+#: model_terms:ir.ui.view,arch_db:eater.Eater Wizard
+msgid "Create a new eater"
+msgstr ""
+
+#. module: eater
+#: model_terms:ir.ui.view,arch_db:eater.beesdoo_partner_form_view
+msgid "Eaters"
+msgstr ""
+
+#. module: eater
+#: model:ir.actions.act_window,name:eater.action_eater_wizard
+#: model_terms:ir.ui.view,arch_db:eater.beesdoo_partner_form_view
+msgid "New Eater"
+msgstr ""
+
+#. module: eater
+#: model:ir.model,name:eater.model_eater_new_wizard
+msgid "eater.new.wizard"
+msgstr ""
+
