+++ eater/readme/DESCRIPTION.rst
@@ -1,0 +1,7 @@
+This module adds the "eater" and "worker" concepts to partners of your participative shop.
+
+- the worker is a voluntary worker of your supermarket
+- the eater, linked to a worker, is usually given the same access rights as the worker
+
+The eater tab is displayed for worker partner.
+A button in this tab allows to create and link an eater to a partner.
