+++ eater/views/partner.xml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+
+
+    <record model="ir.actions.act_window" id="eater_new_wizard_action">
+        <field name="name">New Eater</field>
+        <field name="res_model">eater.new.wizard</field>
+        <field name="view_mode">form</field>
+        <field name="target">new</field>
+    </record>
+
+    <record model="ir.ui.view" id="res_partner_form_view">
+        <field name="name">res.partner.form.view</field>
+        <field name="model">res.partner</field>
+        <field name="inherit_id" ref="base.view_partner_form" />
+        <field name="arch" type="xml">
+
+            <xpath expr="//notebook" position="inside">
+                <page
+                    string="Eaters"
+                    attrs="{'invisible': ['|', ('customer', '=', False), ('eater', 'not in', ('worker_eater'))]}"
+                >
+
+                    <group>
+                        <group>
+                            <field
+                                name="child_eater_ids"
+                                string="Eaters"
+                                widget="many2many_tags"
+                                options="{'no_create': True}"
+                                context="{'default_eater' : 'eater', 'default_customer' : True}"
+                            />
+                        </group>
+                        <group>
+                            <button
+                                string="New Eater"
+                                name="%(action_eater_wizard)d"
+                                type="action"
+                                class="btn btn-primary"
+                            />
+                        </group>
+                    </group>
+
+                </page>
+
+            </xpath>
+
+            <field name="website" position="after">
+                <field name="eater" attrs="{'invisible': [('customer', '=', False)]}" />
+                <field
+                    name="parent_eater_id"
+                    attrs="{'invisible' : [('eater', '!=', 'eater')]}"
+                />
+            </field>
+
+        </field>
+    </record>
+
+</odoo>
