+++ setup/eater/odoo/addons/eater
@@ -1,0 +1,1 @@
+../../../../eater
